# doc/swagger/swagger.yaml
basePath: /api/v1/
definitions:
...
  docs.ErrorBadRequest:
    properties:
      detail:
        example: bad request
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
    type: object
  docs.ErrorInternalServer:
    properties:
      detail:
        example: internal server error
        type: string
      status:
        example: 500
        type: integer
      title:
        example: Internal server error
        type: string
    type: object
...
  docs.ErrorInvalidReservationTimeFrame:
    properties:
      detail:
        example: reservation time frame is invalid
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
    type: object
  docs.ErrorMinimumReservationHours:
    properties:
      detail:
        example: period is shorter than minimun allowed (6 hours)
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
    type: object
  docs.ErrorNotFound:
    properties:
      detail:
        example: not found
        type: string
      status:
        example: 404
        type: integer
      title:
        example: Not Found
        type: string
    type: object
  docs.ErrorReservationNotFound:
    properties:
      detail:
        example: reservation was not found
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
    type: object
...
  docs.ReservationRequest:
    properties:
      car_id:
        example: 0ddac1d8-c7f2-44a6-8c7e-3d06410f7be1
        type: string
      end_date:
        example: "2023-05-16T18:00:00Z"
        type: string
      payment_status:
        example: Paid
        type: string
      start_date:
        example: "2023-05-15T10:00:00Z"
        type: string
      status:
        example: Reserved
        type: string
      user_id:
        example: a29b1af4-9650-4379-8a8b-7f6c4d374e7f
        type: string
    type: object
  docs.ReservationResponse:
    properties:
      car_id:
        example: 0ddac1d8-c7f2-44a6-8c7e-3d06410f7be1
        type: string
      end_date:
        example: "2027-05-22T18:00:00Z"
        type: string
      id:
        example: 882dfcf8-98c9-4a25-9637-ae4564928b10
        type: string
      payment_status:
        example: Paid
        type: string
      start_date:
        example: "2027-05-15T10:00:00Z"
        type: string
      status:
        example: Reserved
        type: string
      user_id:
        example: a29b1af4-9650-4379-8a8b-7f6c4d374e7f
        type: string
    type: object
  docs.Reservations:
    properties:
      reservations:
        items:
          $ref: '#/definitions/docs.ReservationResponse'
        type: array
    type: object
  docs.UserRequest:
    properties:
      email:
        example: isaac.newton@cam.ac.uk
        type: string
      first_name:
        example: Isaac
        type: string
      last_name:
        example: Newton
        type: string
      status:
        example: Active
        type: string
      type:
        example: Customer
        type: string
    type: object
  docs.UserResponse:
    properties:
      email:
        example: isaac.newton@cam.ac.uk
        type: string
      first_name:
        example: Isaac
        type: string
      id:
        example: b6dcf3b3-ec0a-9f31-4379-4b8e7b94a387
        type: string
      last_name:
        example: Newton
        type: string
      status:
        example: Active
        type: string
      type:
        example: Customer
        type: string
    type: object
host: localhost:5050
info:
  contact: {}
  description: This is an API to manage a car rent service
  title: Car Rent API
  version: "1.0"
paths:
...
  /cars/{car_id}/reservations:
    get:
      description: Get reservations by Car id
      operationId: get-reservation-by-car
      parameters:
      - description: Car id
        format: uuid
        in: path
        name: car_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Obtained reservations
          schema:
            $ref: '#/definitions/docs.Reservations'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Get reservations by Car id
      tags:
      - Reservations
...
  /reservations:
    post:
      consumes:
      - application/json
      description: Create a reservation with the provided information
      operationId: create-reservation
      parameters:
      - description: 'Reservation information (allowed statuses: Reserved, Canceled,
          Completed; allowed payment statuses: Paid, Pending, Canceled)'
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/docs.ReservationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created reservation
          schema:
            $ref: '#/definitions/docs.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorMinimumReservationHours'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Create a reservation
      tags:
      - Reservations
  /reservations/{id}:
    delete:
      description: Delete a reservation by UUID
      operationId: delete-reservation
      parameters:
      - description: Reservation UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorReservationNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Delete a reservation
      tags:
      - Reservations
    get:
      description: Get a reservation by UUID
      operationId: get-reservation
      parameters:
      - description: Reservation UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Obtained reservation
          schema:
            $ref: '#/definitions/docs.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorReservationNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Get a reservation
      tags:
      - Reservations
    put:
      consumes:
      - application/json
      description: Update a reservation by UUID
      operationId: update-reservation
      parameters:
      - description: Reservation UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: 'Reservation information (allowed statuses: Reserved, Canceled,
          Completed; allowed payment statuses: Paid, Pending, Canceled)'
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/docs.ReservationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated reservation
          schema:
            $ref: '#/definitions/docs.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorInvalidReservationTimeFrame'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorReservationNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Update a reservation
      tags:
      - Reservations
  /users:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided information
      operationId: register-user
      parameters:
      - description: 'User information (allowed types: Customer, Admin; allowed statuses:
          Active, Inactive)'
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/docs.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created user
          schema:
            $ref: '#/definitions/docs.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Register a new user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete a user by UUID
      operationId: delete-user
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Delete a user
      tags:
      - Users
    get:
      description: Get a user by UUID
      operationId: get-user
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Obtained user
          schema:
            $ref: '#/definitions/docs.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Get a user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update a user by UUID
      operationId: update-user
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: 'User information (allowed types: Customer, Admin; allowed statuses:
          Active, Inactive)'
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/docs.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user
          schema:
            $ref: '#/definitions/docs.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Update a user
      tags:
      - Users
  /users/{user_id}/reservations:
    get:
      description: Get reservations by User id
      operationId: get-reservation-by-user
      parameters:
      - description: User id
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Obtained reservations
          schema:
            $ref: '#/definitions/docs.Reservations'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Get reservations by User id
      tags:
      - Reservations
swagger: "2.0"